put gpsr separate
