+++ horqrux/differentiation/__init__.py
@@ -2,4 +2,4 @@ from __future__ import annotations
 
 from .adjoint import adjoint_expectation
 from .automatic_diff import ad_expectation
-from .shots import finite_shots_fwd
+from .gpsr.shots import finite_shots_fwd
