+++ horqrux/differentiation/gpsr/shots.py
